+++ glade_forms/lib/src/validator/string_validator.dart
@@ -1,16 +1,20 @@
 import 'package:glade_forms/src/core/glade_error_keys.dart';
-import 'package:glade_forms/src/validator/glade_validator.dart';
-import 'package:glade_forms/src/validator/regex_patterns.dart';
-import 'package:glade_forms/src/validator/validator_error/glade_validator_error.dart';
+import 'package:glade_forms/src/validator/validator.dart';
 
 class StringValidator extends GladeValidator<String> {
   StringValidator();
 
   /// Given value can't be empty string (or null).
-  void notEmpty({OnValidateError<String>? devError, Object? key}) => satisfy(
+  void notEmpty({
+    OnValidateError<String>? devError,
+    Object? key,
+    ShouldValidateCallback<String>? shouldValidate,
+  }) =>
+      satisfy(
         (input) => input.isNotEmpty,
         devError: devError ?? (_) => "Value can't be empty",
         key: key ?? GladeErrorKeys.stringEmpty,
+        shouldValidate: shouldValidate,
       );
 
   /// Checks that value is valid email address.
@@ -20,6 +24,7 @@ class StringValidator extends GladeValidator<String> {
     OnValidateError<String>? devError,
     bool allowEmpty = false,
     Object? key,
+    ShouldValidateCallback<String>? shouldValidate,
   }) =>
       satisfy(
         (x) {
@@ -33,6 +38,7 @@ class StringValidator extends GladeValidator<String> {
         },
         devError: devError ?? (value) => 'Value "${value ?? 'NULL'}" is not in e-mail format',
         key: key ?? GladeErrorKeys.stringNotEmail,
+        shouldValidate: shouldValidate,
       );
 
   /// Checks that value is valid URL address.
@@ -44,6 +50,7 @@ class StringValidator extends GladeValidator<String> {
     bool allowEmpty = false,
     bool requiresScheme = false,
     Object key = GladeErrorKeys.stringNotUrl,
+    ShouldValidateCallback<String>? shouldValidate,
   }) =>
       satisfy(
         (value) {
@@ -64,6 +71,7 @@ class StringValidator extends GladeValidator<String> {
         },
         devError: devError ?? (value) => 'Value "${value ?? 'NULL'}" is not valid URL address',
         key: key,
+        shouldValidate: shouldValidate,
       );
 
   /// Matches provided regex [pattern].
@@ -75,6 +83,7 @@ class StringValidator extends GladeValidator<String> {
     bool unicode = false,
     OnValidateError<String>? devError,
     Object? key,
+    ShouldValidateCallback<String>? shouldValidate,
   }) =>
       satisfy(
         (value) {
@@ -85,6 +94,7 @@ class StringValidator extends GladeValidator<String> {
         },
         devError: devError ?? (value) => 'Value "${value ?? 'NULL'}" does not match regex',
         key: key ?? GladeErrorKeys.stringPatternMatch,
+        shouldValidate: shouldValidate,
       );
 
   /// String's length has to be greater or equal to provided [length].
@@ -92,6 +102,7 @@ class StringValidator extends GladeValidator<String> {
     required int length,
     OnValidateError<String>? devError,
     Object? key,
+    ShouldValidateCallback<String>? shouldValidate,
   }) =>
       satisfy(
         (value) {
@@ -99,6 +110,7 @@ class StringValidator extends GladeValidator<String> {
         },
         devError: devError ?? (value) => 'Value "${value ?? 'NULL'}" is shorter than allowed length $length',
         key: key ?? GladeErrorKeys.stringMinLength,
+        shouldValidate: shouldValidate,
       );
 
   /// String's length has to be less or equal to provided [length].
@@ -106,11 +118,13 @@ class StringValidator extends GladeValidator<String> {
     required int length,
     OnValidateError<String>? devError,
     Object? key,
+    ShouldValidateCallback<String>? shouldValidate,
   }) =>
       satisfy(
         (value) => value.length < length,
         devError: devError ?? (value) => 'Value "${value ?? 'NULL'}" is longer than allowed length $length',
         key: key ?? GladeErrorKeys.stringMaxLength,
+        shouldValidate: shouldValidate,
       );
 
   /// String's length has to be equal to provided [length].
@@ -118,10 +132,12 @@ class StringValidator extends GladeValidator<String> {
     required int length,
     OnValidateError<String>? devError,
     Object? key,
+    ShouldValidateCallback<String>? shouldValidate,
   }) =>
       satisfy(
         (value) => value.length == length,
         devError: devError ?? (value) => 'Value "${value ?? 'NULL'}" has to be $length long characters.',
         key: key ?? GladeErrorKeys.stringExactLength,
+        shouldValidate: shouldValidate,
       );
 }