+++ glade_forms/lib/src/validator/validator_error/value_null_error.dart
@@ -3,7 +3,7 @@ import 'package:glade_forms/src/validator/validator_error/glade_validator_error.
 class ValueNullError<T> extends GladeValidatorError<T> {
   ValueNullError({
     required super.value,
+    required super.key,
     OnValidateError<T>? devError,
-    super.key,
   }) : super(devError: devError ?? (_) => "Value can't be null");
 }